requirements.txt
# Core runtime
fastapi>=0.110
uvicorn[standard]>=0.30
streamlit>=1.32

# NLP and modeling
transformers>=4.37
torch>=2.1,<3
accelerate>=0.26
datasets>=2.17
seqeval>=1.2.2

# OCR and imaging
easyocr>=1.7
opencv-python>=4.8
Pillow>=10.0
pyzbar>=0.1.9  # Barcode detection

# Data + utils
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
python-multipart>=0.0.6
tqdm>=4.66
requests>=2.31  # For OpenFoodFacts API

# Jupyter / notebooks
jupyter>=1.0
notebook>=7.0
ipykernel>=6.29
matplotlib>=3.8
seaborn>=0.13

# Optional: ensure Windows compatibility
typing-extensions>=4.9